opts1 is used as uint32_t* and needs same type as opts5

+++ minipro.h
@@ -122,7 +122,7 @@ typedef struct device {
   uint32_t data_memory2_size;
   uint32_t chip_id;  // A vendor-specific chip ID (i.e. 0x1E9502 for ATMEGA48)
   uint8_t chip_id_bytes_count;
-  uint16_t opts1;
+  uint32_t opts1;
   uint16_t opts2;
   uint32_t opts3;
   uint32_t opts4;
